+++ server.cpp
@@ -205,7 +205,7 @@ void Server::getfile(Session& scmd) {
     scmd.sendmsg("BEGIN");
     // 获得文件大小
     fin.seekg(0, fin.end);
-    int fsize = fin.tellg();
+    size_t fsize = fin.tellg();
     fin.seekg(0, fin.beg);
 
     datasession.sendstream(fin, fsize);
